Replace `groupfinder` with `principals_for_userid`
