+++ h/security/__init__.py
@@ -6,4 +6,7 @@ from h.security.encryption import (  # noqa:F401
 )
 from h.security.identity import Identity  # noqa:F401
 from h.security.permissions import Permission  # noqa:F401
-from h.security.principals import principals_for_identity  # noqa:F401
+from h.security.principals import (  # noqa:F401
+    principals_for_identity,
+    principals_for_userid,
+)
